+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 -   **terra-stopwatch** is deprecated now.
 -   **terra-radio-group** & **terra-radio-input** are deprecated now. Use angular material's [radio-button](https://material.angular.io/components/radio/overview) instead.
 -   **terra-slider** is deprecated now. Use angular material's [slider](https://material.angular.io/components/slider/overview) instead.
+-   **file-chooser** has been deprecated. Use material's [button](https://material.angular.io/components/button) and [dialog](https://material.angular.io/components/dialog) in conjunction with our [file-browser](./src/lib/components/file-browser) instead.
 
 # 11.1.0 (23.03.2021)