+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Features
 
 -   **filter** added new components and directives to create a filter for tables. See this [directory](./src/lib/components/filter) for more details.
+
+### Minor changes
+
 -   **tc-filter** has been deprecated.
 
 # 11.2.0 (10.06.2021)